Initial commit to move slackbot to go modules.
