+++ database/database.go
@@ -2,7 +2,7 @@ package database
 
 import (
 	"fmt"
-	"github.com/buildit/slackbot/pkg/config"
+	"github.com/buildit/slackbot/config"
 	log "github.com/sirupsen/logrus"
 	"go.etcd.io/bbolt"
 )
