+++ go.mod
@@ -1,0 +1,23 @@
+module github.com/buildit/slackbot
+
+require (
+	code.cloudfoundry.org/clock v0.0.0-20180518195852-02e53af36e6c // indirect
+	github.com/Microsoft/ApplicationInsights-Go v0.4.2
+	github.com/golang/protobuf v1.3.1 // indirect
+	github.com/gorilla/mux v1.7.1
+	github.com/gorilla/websocket v1.4.0 // indirect
+	github.com/kelseyhightower/envconfig v1.3.0
+	github.com/kr/pretty v0.1.0 // indirect
+	github.com/lusis/go-slackbot v0.0.0-20180109053408-401027ccfef5 // indirect
+	github.com/lusis/slack-test v0.0.0-20190426140909-c40012f20018 // indirect
+	github.com/microsoft/ApplicationInsights-Go v0.4.2
+	github.com/nlopes/slack v0.5.0
+	github.com/onsi/ginkgo v1.8.0 // indirect
+	github.com/onsi/gomega v1.5.0 // indirect
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/satori/go.uuid v1.2.0 // indirect
+	github.com/sirupsen/logrus v1.4.1
+	github.com/tedsuo/ifrit v0.0.0-20180802180643-bea94bb476cc // indirect
+	go.etcd.io/bbolt v1.3.2
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+)
