+++ main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"github.com/microsoft/ApplicationInsights-Go/appinsights"
-	"github.com/buildit/slackbot/pkg/config"
-	"github.com/buildit/slackbot/pkg/database"
-	"github.com/buildit/slackbot/pkg/ai"
-	"github.com/buildit/slackbot/pkg/service"
+	"github.com/Microsoft/ApplicationInsights-Go/appinsights"
+	"github.com/buildit/slackbot/ai"
+	"github.com/buildit/slackbot/config"
+	"github.com/buildit/slackbot/database"
+	"github.com/buildit/slackbot/service"
 	"github.com/gorilla/mux"
 	log "github.com/sirupsen/logrus"
 	"net/http"
